CBIR/requirements.txt
opencv-python>=4.8.0
numpy<2.0,>=1.26.4
faiss-cpu>=1.7.4
scikit-image>=0.21.0
pandas>=2.0.0
scikit-learn>=1.3.0
tensorflow>=2.15.0
streamlit>=1.30.0
streamlit-cropper>=0.2.0
pillow>=10.0.0
